+++ NEWS.md
@@ -56,6 +56,7 @@
 
 # uGMAR 3.2.5
 
+* New exported function: 'condmomentPlot' for further visualization of the model.
 * Yet another bug fix: the predict method incurred an error with G-StMAR models. 
 * Corrected degrees of freedom labels for G-StMAR models in the function profile_logliks.
 * Updated the examples.